image saving refactor

+++ application/classes/Goodies.php
@@ -44,26 +44,34 @@ class Goodies
         return array_pop($args);
     }
 
-    static function image_save($name, $path, $width = false, $height = false)
+    static function images_save($name, $path, $width = false, $height = false)
     {
+        $files = array();
         if (!empty($_FILES[$name]['name'])) {
-            $ext = substr(strrchr($_FILES[$name]['name'], '.'), 1);
-            $filename = time() . '_' . Text::random('alnum', 4);
-            $image_file = substr(strrchr(Upload::save($_FILES[$name], $filename . '_tmp.' . $ext, stripcslashes($path) . '/'), DIRECTORY_SEPARATOR), 1);
-            $image = Image::factory(stripcslashes($path) . '/' . $image_file);
-            if ($width && ($image->width > $width)) {
-                $image->resize($width, $image->height);
-            }
-            if ($height && ($image->height > $height)) {
-                $image->resize($image->width, $height);
-            }
-            $image->save();
+            foreach ($_FILES[$name]['error'] as $key => $error) {
+                if ($error == UPLOAD_ERR_OK) {
+                    $ext = substr(strrchr($_FILES[$name]['name'][$key], '.'), 1);
+                    $filename = time() . '_' . Text::random('alnum', 4);
+                    $tmp_name = $_FILES['image']['tmp_name'][$key];
+                    $image_file = $filename . '_tmp.' . $ext;
+                    move_uploaded_file($tmp_name, stripcslashes($path) . '/' . $image_file);
+                    $image = Image::factory(stripcslashes($path) . '/' . $image_file);
+                    if ($width && ($image->width > $width)) {
+                        $image->resize($width, $image->height);
+                    }
+                    if ($height && ($image->height > $height)) {
+                        $image->resize($image->width, $height);
+                    }
+                    $image->save();
 
-            copy(stripcslashes($path) . '/' . $image_file, stripcslashes($path) . '/' . $filename . '.' . $ext);
-            unlink(stripcslashes($path) . '/' . $image_file);
+                    copy(stripcslashes($path) . '/' . $image_file, stripcslashes($path) . '/' . $filename . '.' . $ext);
+                    unlink(stripcslashes($path) . '/' . $image_file);
 
-            return $filename . '.' . $ext;
-        } else return false;
+                    $files[] =  $filename . '.' . $ext;
+                }
+            }
+        return $files;
+        }
+        else return false;
     }
-
 }
